Move finite matmul imports to consumers
